% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_samples}
\alias{gen_samples}
\title{Generate samples based on the decomposition of Y}
\usage{
gen_samples(orthdir, dir, this_sd, nrSample = 1000,
  sampFun = function(n) rnorm(n, mean = 0, sd = this_sd), checkFun)
}
\arguments{
\item{orthdir}{nx1 vector; component orthogonal to the direction of interest}

\item{dir}{nx1 vector; component in the direction of interest}

\item{this_sd}{scalar; (estimated / surrogate) value for the root of error variance}

\item{nrSample}{integer; number of samples to be used}

\item{sampFun}{function; function to generate samples from}

\item{checkFun}{function; function to congruency with initial selection}
}
\description{
Generate samples based on the decomposition of Y
}
