% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfmade.R
\name{print.selfmade}
\alias{print.selfmade}
\title{Generic methods for selfmade objects}
\usage{
\method{print}{selfmade}(x, ...)
}
\arguments{
\item{x}{selfmade object}

\item{...}{further arguments, currently unused.}
}
\description{
Generic methods which can be used for objects 
fitted with the \code{mocasin} function
}
