% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selinf}
\alias{selinf}
\title{Calculate inference based on survived samples and weights}
\usage{
selinf(survr, tstat, w, var_est, alpha)
}
\arguments{
\item{survr}{vector of survived samples}

\item{tstat}{the original test statistic}

\item{w}{weights for the survived samples}

\item{var_est}{(estimated) variance of the test statistic}

\item{alpha}{see \code{\link{mocasin}}}
}
\description{
Calculate inference based on survived samples and weights
}
