% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testvec_for_gamm4}
\alias{testvec_for_gamm4}
\title{Function to calculate the test vector for an object fitted with \code{gamm4}}
\usage{
testvec_for_gamm4(mod, name, sigma2 = NULL, nrlocs = 7)
}
\arguments{
\item{mod}{an object fitted with \code{gamm4}}

\item{name}{character; name of the covariate for which inference should be
calculated}

\item{sigma2}{variance to be used in the covariance definition. If \code{NULL},
the estimate \code{mod$gam$sig2} is used.}

\item{nrlocs}{number of locations at which p-values and intervals are to be computed 
for non-linear terms. This directly corresponds to a sequence of \code{nrlocs} quantiles 
of the given covariate values.}
}
\description{
Function to calculate the test vector for an object fitted with \code{gamm4}
}
\details{
Function provides the test vectors for every location of the given covariate
}
