% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testvec_for_mm}
\alias{testvec_for_mm}
\title{Function to calculate the test vector for an \code{merMod} object}
\usage{
testvec_for_mm(mod, which = NULL, marginal = TRUE, VCOV = NULL,
  G = NULL, sig2 = NULL, efficient = TRUE)
}
\arguments{
\item{mod}{\code{merMod} object}

\item{which}{integer; if NULL, test vector is created for all 
coefficients, otherwise for the which'th coefficient}

\item{marginal}{logical; whether to construct a test vector
from the marginal or condition perspective}

\item{VCOV}{covariance used for the construction of test vector}

\item{G}{true random effect covariance}

\item{sig2}{true error variance}

\item{efficient}{logical; whether to compute the 
test vector corresponding to the efficient beta estimator
in the marginal case}
}
\description{
Function to calculate the test vector for an \code{merMod} object
}
\details{
Function provides 2 different marginal test vectors
(efficient = TRUE / FALSE) and one conditional test vector

Note that the covariance of residuals R is assumed to be diagonal.
}
