% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pval_vT_cov}
\alias{pval_vT_cov}
\title{Calculate selective p-value for given covariance}
\usage{
pval_vT_cov(vT, VCOV, this_y, nrSamples, checkFun, twosided = TRUE,
  bayesian = FALSE, alpha = 0.05, ...)
}
\arguments{
\item{vT}{test vector of function}

\item{VCOV}{covariance used for distribution of test statistic}

\item{this_y}{original response vector}

\item{nrSamples}{number of samples to be used}

\item{checkFun}{function; function to congruency with initial selection}

\item{twosided}{logical; compute two-sided p-values?}

\item{...}{further arguments passed to vT if specified as function}

\item{coefNr}{which coefficient p-values should be calculated for}

\item{bayesian_sd}{if not NULL, this variance is used as variance of the test statistic.
The idea of this argument is to allow for usage of a bayesian covariance, but 
this argument can also be used otherwise.}
}
\description{
Calculate selective p-value for given covariance
}
